refactor: split log store into multiple files
